Enum values are now `i64`. Additionally, they are now stored in a map, which allows arbitrary values, although the syntax doesn't currently support assigning values to enum variants.

+++ crates/aranya-policy-module/src/data.rs
@@ -110,7 +110,7 @@ pub enum Value {
     /// A unique identifier.
     Id(Id),
     /// Enumeration value
-    Enum(String, usize),
+    Enum(String, i64),
     /// Empty optional value
     None,
 }
